+++ pkg/cuvs/ivf_pq_test.go
@@ -436,7 +436,7 @@ func TestGpuIvfPqExtend(t *testing.T) {
 		t.Fatalf("Extend failed: %v", err)
 	}
 
-	if got := index.Len(); got != uint32(nBase+nExt) {
+	if got := index.Len(); got != uint64(nBase+nExt) {
 		t.Errorf("Len() = %d, want %d", got, nBase+nExt)
 	}
 
@@ -510,7 +510,7 @@ func TestGpuIvfPqExtendFloat(t *testing.T) {
 		t.Fatalf("ExtendFloat failed: %v", err)
 	}
 
-	if got := index.Len(); got != uint32(nBase+nExt) {
+	if got := index.Len(); got != uint64(nBase+nExt) {
 		t.Errorf("Len() = %d, want %d", got, nBase+nExt)
 	}
